Fix repeated unnecessary processing of attributes
As a shortcut when an attribute was changed the code would process every
eligible attribute on the element. This would cause issues if the
replacement was "recursive" like e -> ee as a change in one attribute
would lead to another attribute's number of E's doubling. Now all
attribute mutations are collected and pulled out into a separate array
and processed after the other mutation types.